+++ piccolo/query/methods/objects.py
@@ -1,13 +1,13 @@
 from __future__ import annotations
 
 import typing as t
-from dataclasses import dataclass
 
 from piccolo.columns.column_types import ForeignKey
 from piccolo.columns.combination import And, Where
-from piccolo.custom_types import Combinable
+from piccolo.custom_types import Combinable, TableInstance
 from piccolo.engine.base import Batch
 from piccolo.query.base import Query
+from piccolo.query.methods.select import Select
 from piccolo.query.mixins import (
     AsOfDelegate,
     CallbackDelegate,
@@ -20,30 +20,49 @@ from piccolo.query.mixins import (
     PrefetchDelegate,
     WhereDelegate,
 )
+from piccolo.query.proxy import Proxy
 from piccolo.querystring import QueryString
 from piccolo.utils.dictionary import make_nested
 from piccolo.utils.sync import run_sync
 
-from .select import Select
-
 if t.TYPE_CHECKING:  # pragma: no cover
     from piccolo.columns import Column
-    from piccolo.table import Table
 
 
-@dataclass
-class GetOrCreate:
-    query: Objects
-    where: Combinable
-    defaults: t.Dict[t.Union[Column, str], t.Any]
+###############################################################################
 
-    async def run(self):
-        instance = await self.query.get(self.where).run()
+
+class GetOrCreate(
+    Proxy["Objects[TableInstance]", TableInstance], t.Generic[TableInstance]
+):
+    def __init__(
+        self,
+        query: Objects[TableInstance],
+        table_class: t.Type[TableInstance],
+        where: Combinable,
+        defaults: t.Dict[Column, t.Any],
+    ):
+        self.query = query
+        self.table_class = table_class
+        self.where = where
+        self.defaults = defaults
+
+    async def run(
+        self, node: t.Optional[str] = None, in_pool: bool = True
+    ) -> TableInstance:
+        """
+        :raises ValueError:
+            If more than one matching row is found.
+
+        """
+        instance = await self.query.get(self.where).run(
+            node=node, in_pool=in_pool
+        )
         if instance:
             instance._was_created = False
             return instance
 
-        instance = self.query.table()
+        instance = self.table_class(_data=self.defaults)
 
         # If it's a complex `where`, there can be several column values to
         # extract e.g. (Band.name == 'Pythonistas') & (Band.popularity == 1000)
@@ -60,12 +79,7 @@ class GetOrCreate:
                     # to this table.
                     setattr(instance, column._meta.name, value)
 
-        for column, value in self.defaults.items():
-            if isinstance(column, str):
-                column = instance._meta.get_column_by_name(column)
-            setattr(instance, column._meta.name, value)
-
-        await instance.save().run()
+        await instance.save().run(node=node, in_pool=in_pool)
 
         # If the user wants us to prefetch related objects, for example:
         #
@@ -85,46 +99,86 @@ class GetOrCreate:
                 .run()
             )
 
+        instance = t.cast(TableInstance, instance)
         instance._was_created = True
         return instance
 
-    def __await__(self):
-        """
-        If the user doesn't explicity call .run(), proxy to it as a
-        convenience.
-        """
-        return self.run().__await__()
 
-    def run_sync(self):
-        return run_sync(self.run())
-
-    def prefetch(self, *fk_columns) -> GetOrCreate:
-        self.query.prefetch(*fk_columns)
-        return self
+class Get(
+    Proxy["First[TableInstance]", t.Optional[TableInstance]],
+    t.Generic[TableInstance],
+):
+    pass
 
 
-@dataclass
-class Create:
-    table_class: t.Type[Table]
-    columns: t.Dict[str, t.Any]
+class First(
+    Proxy["Objects[TableInstance]", t.Optional[TableInstance]],
+    t.Generic[TableInstance],
+):
+    async def run(
+        self, node: t.Optional[str] = None, in_pool: bool = True
+    ) -> t.Optional[TableInstance]:
+        objects = await self.query.run(
+            node=node, in_pool=in_pool, use_callbacks=False
+        )
 
-    async def run(self):
+        results = objects[0] if objects else None
+
+        modified_response: t.Optional[
+            TableInstance
+        ] = await self.query.callback_delegate.invoke(
+            results=results, kind=CallbackType.success
+        )
+        return modified_response
+
+
+class Create(t.Generic[TableInstance]):
+    """
+    This is provided as a simple convenience. Rather than running::
+
+        band = Band(name='Pythonistas')
+        await band.save()
+
+    We can instead do it in a single line::
+
+        band = Band.objects().create(name='Pythonistas')
+
+    """
+
+    def __init__(
+        self,
+        table_class: t.Type[TableInstance],
+        columns: t.Dict[str, t.Any],
+    ):
+        self.table_class = table_class
+        self.columns = columns
+
+    async def run(
+        self,
+        node: t.Optional[str] = None,
+        in_pool: bool = True,
+    ) -> TableInstance:
         instance = self.table_class(**self.columns)
-        await instance.save().run()
+        await instance.save().run(node=node, in_pool=in_pool)
         return instance
 
-    def __await__(self):
+    def __await__(self) -> t.Generator[None, None, TableInstance]:
         """
         If the user doesn't explicity call .run(), proxy to it as a
         convenience.
         """
         return self.run().__await__()
 
-    def run_sync(self):
-        return run_sync(self.run())
+    def run_sync(self, *args, **kwargs) -> TableInstance:
+        return run_sync(self.run(*args, **kwargs))
 
 
-class Objects(Query):
+###############################################################################
+
+
+class Objects(
+    Query[TableInstance, t.List[TableInstance]], t.Generic[TableInstance]
+):
     """
     Almost identical to select, except you have to select all fields, and
     table instances are returned, rather than just data.
@@ -144,7 +198,7 @@ class Objects(Query):
 
     def __init__(
         self,
-        table: t.Type[Table],
+        table: t.Type[TableInstance],
         prefetch: t.Sequence[t.Union[ForeignKey, t.List[ForeignKey]]] = (),
         **kwargs,
     ):
@@ -160,18 +214,18 @@ class Objects(Query):
         self.prefetch(*prefetch)
         self.where_delegate = WhereDelegate()
 
-    def output(self, load_json: bool = False) -> Objects:
+    def output(self: Self, load_json: bool = False) -> Self:
         self.output_delegate.output(
             as_list=False, as_json=False, load_json=load_json
         )
         return self
 
     def callback(
-        self,
+        self: Self,
         callbacks: t.Union[t.Callable, t.List[t.Callable]],
         *,
         on: CallbackType = CallbackType.success,
-    ) -> Objects:
+    ) -> Self:
         self.callback_delegate.callback(callbacks, on=on)
         return self
 
@@ -179,46 +233,23 @@ class Objects(Query):
         self.as_of_delegate.as_of(interval)
         return self
 
-    def limit(self, number: int) -> Objects:
+    def limit(self: Self, number: int) -> Self:
         self.limit_delegate.limit(number)
         return self
 
-    def first(self) -> Objects:
-        self.limit_delegate.first()
-        return self
-
     def prefetch(
-        self, *fk_columns: t.Union[ForeignKey, t.List[ForeignKey]]
-    ) -> Objects:
+        self: Self, *fk_columns: t.Union[ForeignKey, t.List[ForeignKey]]
+    ) -> Self:
         self.prefetch_delegate.prefetch(*fk_columns)
         return self
 
-    def get(self, where: Combinable) -> Objects:
-        self.where_delegate.where(where)
-        self.limit_delegate.first()
-        return self
-
-    def offset(self, number: int) -> Objects:
+    def offset(self: Self, number: int) -> Self:
         self.offset_delegate.offset(number)
         return self
 
-    def get_or_create(
-        self,
-        where: Combinable,
-        defaults: t.Dict[t.Union[Column, str], t.Any] = None,
-    ):
-        if defaults is None:
-            defaults = {}
-        return GetOrCreate(query=self, where=where, defaults=defaults)
-
-    def create(self, **columns: t.Any):
-        return Create(table_class=self.table, columns=columns)
-
     def order_by(
-        self,
-        *columns: t.Union[Column, str, OrderByRaw],
-        ascending: bool = True,
-    ) -> Objects:
+        self: Self, *columns: t.Union[Column, str, OrderByRaw], ascending=True
+    ) -> Self:
         _columns: t.List[t.Union[Column, OrderByRaw]] = []
         for column in columns:
             if isinstance(column, str):
@@ -229,12 +260,39 @@ class Objects(Query):
         self.order_by_delegate.order_by(*_columns, ascending=ascending)
         return self
 
-    def where(self, *where: Combinable) -> Objects:
+    def where(self: Self, *where: Combinable) -> Self:
         self.where_delegate.where(*where)
         return self
 
+    ###########################################################################
+
+    def first(self: Self) -> First[TableInstance]:
+        self.limit_delegate.limit(1)
+        return First[TableInstance](query=self)
+
+    def get(self: Self, where: Combinable) -> Get[TableInstance]:
+        self.where_delegate.where(where)
+        self.limit_delegate.limit(1)
+        return Get[TableInstance](query=First[TableInstance](query=self))
+
+    def get_or_create(
+        self: Self,
+        where: Combinable,
+        defaults: t.Dict[Column, t.Any] = None,
+    ) -> GetOrCreate[TableInstance]:
+        if defaults is None:
+            defaults = {}
+        return GetOrCreate[TableInstance](
+            query=self, table_class=self.table, where=where, defaults=defaults
+        )
+
+    def create(self: Self, **columns: t.Any) -> Create[TableInstance]:
+        return Create[TableInstance](table_class=self.table, columns=columns)
+
+    ###########################################################################
+
     async def batch(
-        self,
+        self: Self,
         batch_size: t.Optional[int] = None,
         node: t.Optional[str] = None,
         **kwargs,
@@ -246,14 +304,7 @@ class Objects(Query):
         return await self.table._meta.db.batch(self, **kwargs)
 
     async def response_handler(self, response):
-        if self.limit_delegate._first:
-            if len(response) == 0:
-                return None
-            if self.output_delegate._output.nested:
-                return make_nested(response[0])
-            else:
-                return response[0]
-        elif self.output_delegate._output.nested:
+        if self.output_delegate._output.nested:
             return [make_nested(i) for i in response]
         else:
             return response
@@ -287,3 +338,34 @@ class Objects(Query):
             select.output_delegate.output(nested=True)
 
         return select.querystrings
+
+    ###########################################################################
+
+    async def run(
+        self,
+        node: t.Optional[str] = None,
+        in_pool: bool = True,
+        use_callbacks: bool = True,
+    ) -> t.List[TableInstance]:
+        results = await super().run(node=node, in_pool=in_pool)
+
+        if use_callbacks:
+            # With callbacks, the user can return any data that they want.
+            # Assume that most of the time they will still return a list of
+            # Table instances.
+            modified: t.List[
+                TableInstance
+            ] = await self.callback_delegate.invoke(
+                results, kind=CallbackType.success
+            )
+            return modified
+        else:
+            return results
+
+    def __await__(
+        self,
+    ) -> t.Generator[None, None, t.List[TableInstance]]:
+        return super().__await__()
+
+
+Self = t.TypeVar("Self", bound=Objects)